Fix incorrect type.
git-svn-id: https://svn.apache.org/repos/asf/incubator/lucy/trunk@1222511 13f79535-47bb-0310-9956-ffa450edef68

+++ core/Lucy/Index/Posting/RichPosting.c
@@ -89,7 +89,7 @@ RichPost_read_record(RichPosting *self, InStream *instream) {
             = (float*)REALLOCATE(self->prox_boosts, num_prox * sizeof(float));
     }
     uint32_t *positions    = self->prox;
-    uint32_t *prox_boosts  = self->prox_boosts;
+    float    *prox_boosts  = self->prox_boosts;
 
     while (num_prox--) {
         position += InStream_Read_C32(instream);
